Small fix for 64 bit on iOS.

+++ TICoreDataSync/04 Utilities/FZACryptor/FZAKeyManageriPhone.m
@@ -119,7 +119,7 @@
         if( searchResult == errSecItemNotFound ) {
             TICDSLog(TICDSLogVerbosityEveryStep, @"FZACryptor iOS Key Manager didn't find a stored key");
         } else {
-            TICDSLog(TICDSLogVerbosityErrorsOnly, @"FZACryptor iOS Key Manager search error: %ld", searchResult);
+            TICDSLog(TICDSLogVerbosityErrorsOnly, @"FZACryptor iOS Key Manager search error: %d", (int)searchResult);
         }
     }
     return [theKey autorelease];
@@ -132,7 +132,7 @@
     if( clearResult == errSecSuccess ) {
         TICDSLog(TICDSLogVerbosityEveryStep, @"FZACryptor iOS Key Manager deleted keychain items");
     } else {
-        TICDSLog(TICDSLogVerbosityErrorsOnly, @"FZACryptor iOS Key Manager failed to delete keychain items: %ld", clearResult);
+        TICDSLog(TICDSLogVerbosityErrorsOnly, @"FZACryptor iOS Key Manager failed to delete keychain items: %d", (int)clearResult);
     }
 }
